+++ pwmanager-app/dokumentaatio/changelog.md
@@ -43,3 +43,4 @@
 - PasswordsView luokkaan luotu metodi _handle_generate_password joka autogeneroi 16 characterin random salasanan
 - PasswordListView luokkaan luotu metodi _copy_to_clipboard_handler joka kopioi halutun salasanan clipboardiin
 - Testi lisätty että salasanojen hakeminen tietyllä käyttäjällä palauttaa kaikki käyttäjän salasanat
+- Testi lisätty että UserService palauttaa sisäänkirjautuneen käyttäjän tiedot metodilla get_current_user